requirements - Copy.txt
Babel==2.2.0
Click==7.0
Delorean==0.6.0
Django==1.9.5
Werkzeug==0.11.5
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
chardet==2.3.0
colorama==0.3.2
django-appconf==1.0.1
django-cachalot==1.4.1
django-celery==3.1.17
django-compressor==2.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-jet==0.1.4
django-reversion==2.0.6
docutils>=0.12
easysnmp==0.2.4
feedparser==5.2.1
html5lib==0.999
humanize==0.5.1
influxdb==3.0.0
kafka-python==1.4.6
kombu==3.0.35
lesscpy==0.11.1
mysqlclient==1.3.7
numpy==1.11.0
ply==3.8
prompt-toolkit==2.0.10
pygobject>=3.14.0
pylibmc==1.5.1
pymodbus==2.1.0
pyserial==3.4
python-apt>=0.9.3.12
python-dateutil==2.5.2
pytz==2016.3
rcssmin==1.0.6
requests==2.4.3
rjsmin==1.0.12
sh==1.11
six==1.11.0
sqlparse==0.1.19
template-utils==0.4p2
threadpool==1.3.2
tzlocal==1.2.2
uModbus==1.0.2
uWSGI==2.0.12
urllib3==1.9.1
wcwidth==0.1.7
wheel==0.24.0
